Prepare core for Switch and StateMachine nodes
